Fix build with gcc4 on x86-64: ignore errors when casting pointers to int from regs & fpu members + code cache. This is possible because data is allocated in 32-bit space and we force allocation of translation cache to those bounds too.

+++ BasiliskII/src/uae_cpu/compiler/codegen_x86.h
@@ -243,10 +243,10 @@ typedef unsigned short	_us;
 typedef signed int	_sl;
 typedef unsigned int	_ul;
 
-#define _UC(X)		((_uc  )(X))
-#define _US(X)		((_us  )(X))
-#define _SL(X)		((_sl  )(X))
-#define _UL(X)		((_ul  )(X))
+#define _UC(X)		((_uc  )(unsigned long)(X))
+#define _US(X)		((_us  )(unsigned long)(X))
+#define _SL(X)		((_sl  )(unsigned long)(X))
+#define _UL(X)		((_ul  )(unsigned long)(X))
 
 # define _PUC(X)	((_uc *)(X))
 # define _PUS(X)	((_us *)(X))
